Fixed GCC warnings
Summary:

Resolved T21650: removed unneeded classes, structs and typedefs in rttbBase.h


replaced stringstream by C++11 to_string()


Resolved T21650: removed unneeded classes, structs and typedefs in rttbBase.h


Merge branch 'T21650-overheadBaseType'

Removed unused functions


Added testing functions for DoseIteratorInterface


Fixed typo in function name


Added tests for setStructureLabelFilterActive(true)


Added test for getGridSize() and isGridHomogeneous()


Optimized code coverage


Cleaned StructureSet, removed default constructor


Added Test for StructureSet


DVH comparator also checks values.


Increased code coverage of DVHCalculator


Added testing for baseType


Added tests to increase code coverage


Additional tests for GenericDoseIterator and GenericMaskedDoseIterator with added inhomogeneous grid DummyDoseAccessor.


Improved coverage of MaskVoxel


Merge branch 'T22863-optimizzeCoreCodeCoverage'

Merge branch 'master' of https://phabricator.mitk.org/source/rttb

Fixed warnings for GCC compiler.


fixed typo


Added header file for rttbUtils


Prevent same file name of DoseTool output.


Removed further warnings by GCC.


Merge branch 'master' of https://phabricator.mitk.org/source/rttb

Merge branch 'T23158-removeGCCWarnings'

Test Plan: tests run fine

Reviewers: reicht

Reviewed By: reicht

Maniphest Tasks: T21650

Differential Revision: https://phabricator.mitk.org/D65

+++ code/algorithms/rttbMOHxVolumeToDoseMeasureCollectionCalculator.cpp
@@ -45,7 +45,7 @@ namespace rttb
 				double countVoxels = 0;
 				double sum = 0;
 
-				for (auto i = _doseVector.size() - 1; i != -1; i--)
+				for (auto i = _doseVector.size() - 1; i!=static_cast<size_t>(-1); i--)
 				{
 					double voxelProportion = _voxelProportionVector.at(i);
 					countVoxels += voxelProportion;
